Изменил сущности Person, Customer, Employee, наследования и репозитории
