--- src/main/java/com/example/team5animalsheltertelegrambot/entity/Customer.java
@@ -1,16 +1,0 @@
-package com.example.team5animalsheltertelegrambot.entity;
-
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@Entity
-@NoArgsConstructor
-@Table(name = "ast_customer")
-public class Customer extends PersonEntity {
-    public Customer(String firstName, String lastName) {
-        setFirstName(firstName);
-        setLastName(lastName);
-    }
-}
