--- src/main/java/com/example/team5animalsheltertelegrambot/entity/PersonEntity.java
@@ -1,60 +1,0 @@
-package com.example.team5animalsheltertelegrambot.entity;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import javax.validation.constraints.NotEmpty;
-
-@MappedSuperclass
-public class PersonEntity extends BaseEntity{
-
-    @NotEmpty
-    @Column(name = "first_name")
-    private String firstName;
-
-    @NotEmpty
-    @Column(name = "last_name")
-    private String lastName;
-
-    @Column(name = "phone")
-    private String phone;
-
-    @Column(name = "description")
-    private String description;
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @Override
-    public String toString() {
-        return getLastName() + " " + getFirstName();
-    }
-}
