--- src/main/java/com/example/team5animalsheltertelegrambot/repository/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.example.team5animalsheltertelegrambot.repository;
-
-import com.example.team5animalsheltertelegrambot.entity.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
-}
