+++ src/main/java/com/example/team5animalsheltertelegrambot/repository/PersonRepository.java
@@ -1,9 +1,9 @@
 package com.example.team5animalsheltertelegrambot.repository;
 
-import com.example.team5animalsheltertelegrambot.entity.Employee;
+import com.example.team5animalsheltertelegrambot.entity.person.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface PersonRepository<T extends Person> extends JpaRepository<T, Integer> {
 }
